requirements.txt
Flask==2.3.2
PyJWT==2.7.0
msal==1.22.0
requests==2.30.0
